+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/EpisodeManager.kt
@@ -133,5 +133,7 @@ interface EpisodeManager {
 
     suspend fun updateDownloadUrl(episode: PodcastEpisode): String?
 
+    suspend fun updateThumbnailStatusIfUnknown(episodeUuid: String, thumbnailStatus: Int)
+
     suspend fun getAllPodcastEpisodes(pageLimit: Int): Flow<Pair<PodcastEpisode, Int>>
 }